Deprecate omc4py.compiler
